+++ waspc/src/Wasp/Validator.hs
@@ -85,6 +85,17 @@ eqJust expected (Just actual)
 eqJust expected Nothing =
   failure $ "Missing value, expected " ++ show expected ++ "."
 
+-- | Validates that the list value exists and contains all of the required elements.
+-- Extra elements are allowed.
+containsAllJust :: (Eq a, Show a) => [a] -> Validator (Maybe [a])
+containsAllJust requiredElements (Just actual)
+  | null missing = success
+  | otherwise = failure $ "Missing required values: " ++ show missing ++ "."
+  where
+    missing = filter (`notElem` actual) requiredElements
+containsAllJust requiredElements Nothing =
+  failure $ "Missing value, expected at least " ++ show requiredElements ++ "."
+
 -- | Runs the inner validator only if the value is Just. If the value is
 -- Nothing, no validation is needed and the validation succeeds.
 ifJust :: Validator a -> Validator (Maybe a)